acpi: make aml build routines general
The virt/i386 acpi tables also use those routines and move
those into aml-build.c for general usage.
